refactor(notebook): share host-neutral surface state
