+++ apps/notebook/src/contexts/PresenceContext.tsx
@@ -1,50 +1,24 @@
-/**
- * React context for presence functionality.
- *
- * Provides cursor/selection sending methods to the component tree,
- * wrapping the usePresence hook to avoid prop drilling.
- */
-
+import type { ReactNode } from "react";
+import { useCallback, useMemo } from "react";
 import type { NotebookInteractionTarget } from "runtimed";
-import { createContext, type ReactNode, useCallback, useContext, useMemo } from "react";
+import {
+  PresenceValueProvider,
+  type PresenceContextValue,
+} from "@/components/notebook/presence-context";
 import { usePresence } from "../hooks/usePresence";
 
-export interface PresenceContextValue {
-  /** Send cursor position for a cell */
-  setCursor: (cellId: string, line: number, column: number) => void;
-  /** Send selection range for a cell */
-  setSelection: (
-    cellId: string,
-    anchorLine: number,
-    anchorCol: number,
-    headLine: number,
-    headCol: number,
-  ) => void;
-  /** Send cell-level focus (no cursor position) */
-  setFocus: (cellId: string) => void;
-  /** Send active notebook interaction target */
-  setInteraction: (target: NotebookInteractionTarget) => void;
-  /** The local peer's ID */
-  peerId: string | null;
-}
-
-const PresenceContext = createContext<PresenceContextValue | null>(null);
+export {
+  PresenceValueProvider,
+  usePresenceContext,
+  usePresenceContextRequired,
+  type PresenceContextValue,
+} from "@/components/notebook/presence-context";
 
 interface PresenceProviderProps {
   peerId: string;
   peerLabel?: string;
   actorLabel?: string;
   children: ReactNode;
-}
-
-export function PresenceValueProvider({
-  value,
-  children,
-}: {
-  value: PresenceContextValue | null;
-  children: ReactNode;
-}) {
-  return <PresenceContext.Provider value={value}>{children}</PresenceContext.Provider>;
 }
 
 export function PresenceProvider({
@@ -94,25 +68,5 @@ export function PresenceProvider({
     [setCursor, setSelection, setFocus, setInteraction, peerId],
   );
 
-  return <PresenceContext.Provider value={value}>{children}</PresenceContext.Provider>;
-}
-
-/**
- * Hook to access presence context.
- * Returns null if used outside a PresenceProvider.
- */
-export function usePresenceContext(): PresenceContextValue | null {
-  return useContext(PresenceContext);
-}
-
-/**
- * Hook that throws if presence context is not available.
- * Use this when you need presence and it's an error if it's missing.
- */
-export function usePresenceContextRequired(): PresenceContextValue {
-  const ctx = useContext(PresenceContext);
-  if (!ctx) {
-    throw new Error("usePresenceContextRequired must be used within a PresenceProvider");
-  }
-  return ctx;
+  return <PresenceValueProvider value={value}>{children}</PresenceValueProvider>;
 }
