+++ apps/notebook/src/notebook-surface.ts
@@ -2,11 +2,4 @@ export { CodeCell, type HiddenGroupCellSummary } from "./components/CodeCell";
 export { MarkdownCell } from "./components/MarkdownCell";
 export { NotebookView, type NotebookViewProps } from "./components/NotebookView";
 export { RawCell } from "./components/RawCell";
-export { PresenceValueProvider, type PresenceContextValue } from "./contexts/PresenceContext";
-export { CrdtBridgeProvider } from "./hooks/useCrdtBridge";
-export { createNotebookController } from "./lib/notebook-controller";
-export { createNotebookCellId, type NotebookCellIdRandomSource } from "./lib/notebook-cell-id";
 export { materializeChangeset } from "./lib/frame-pipeline";
-export { resetPoolState, setPoolState } from "./lib/pool-state";
-export { startCursorDispatch } from "./lib/cursor-registry";
-export { emitBroadcast, emitPresence } from "./lib/notebook-frame-bus";
