+++ src/components/notebook/__tests__/projection-lifecycle.test.ts
@@ -1,40 +1,47 @@
 import assert from "node:assert/strict";
-import { afterEach, describe, it } from "node:test";
-import { getCellById, updateCellSourceById } from "@/components/notebook/state/cell-store";
+import { afterEach, describe, it } from "vite-plus/test";
+import {
+  getCellById,
+  getCellIdsSnapshot,
+  updateCellSourceById,
+} from "@/components/notebook/state/cell-store";
 import {
   getCellExecutionId,
   getExecutionById,
   getNotebookQueueProjection,
+  resetNotebookExecutions,
   setExecution,
   setNotebookQueueProjection,
 } from "@/components/notebook/state/execution-store";
-import { getOutputById } from "@/components/notebook/state/output-store";
+import { getOutputById, resetNotebookOutputs } from "@/components/notebook/state/output-store";
+import {
+  cleanupNotebookProjectionForRemovedCells,
+  projectNotebookCellsIntoViewStores,
+  resetNotebookProjectionUnlessPreserved,
+  resetNotebookViewStoreProjection,
+} from "@/components/notebook/state/projection-lifecycle";
 import { createNotebookViewModelFromNotebookCells } from "@/components/notebook/state/view-model-store";
 import { setMarkdownProjectionProjector } from "@/lib/markdown-projection";
-import {
-  cleanupCloudProjectionForRemovedCells,
-  projectCloudCellsIntoNotebookViewStores,
-  resetCloudViewStoreProjection,
-} from "../viewer/notebook-view-store-bridge.ts";
 
-describe("cloud NotebookView store bridge", () => {
+describe("notebook projection lifecycle", () => {
   let restoreMarkdownProjectionProjector: (() => void) | undefined;
 
   afterEach(() => {
     restoreMarkdownProjectionProjector?.();
     restoreMarkdownProjectionProjector = undefined;
-    resetCloudViewStoreProjection();
+    resetNotebookViewStoreProjection();
+    resetNotebookExecutions();
+    resetNotebookOutputs();
   });
 
   it("keeps outline projection tied to the shared NotebookView store after local source edits", () => {
     restoreMarkdownProjectionProjector = setMarkdownProjectionProjector(testMarkdownProjector);
 
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       {
         id: "intro",
         cellType: "markdown",
         source: "# Original title",
-        language: null,
         executionId: null,
         executionCount: null,
         outputs: [],
@@ -60,15 +67,14 @@ describe("cloud NotebookView store bridge", () => {
       id: "intro",
       cellType: "markdown" as const,
       source: "# Stable title",
-      language: null,
       executionId: null,
       executionCount: null,
       outputs: [],
       metadata: {},
     };
 
-    projectCloudCellsIntoNotebookViewStores([cell]);
-    projectCloudCellsIntoNotebookViewStores([cell]);
+    projectNotebookCellsIntoViewStores([cell]);
+    projectNotebookCellsIntoViewStores([cell]);
 
     assert.equal(projectCalls, 1);
     assert.deepEqual(outlineTitlesFromStore(), ["Stable title"]);
@@ -80,12 +86,11 @@ describe("cloud NotebookView store bridge", () => {
       queued_cell_ids: ["cell-queued"],
     });
 
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       {
         id: "cell-running",
         cellType: "code",
         source: "print('running')",
-        language: "python",
         executionId: "exec-running",
         executionCount: null,
         outputs: [],
@@ -95,7 +100,6 @@ describe("cloud NotebookView store bridge", () => {
         id: "cell-queued",
         cellType: "code",
         source: "print('queued')",
-        language: "python",
         executionId: "exec-queued",
         executionCount: null,
         outputs: [],
@@ -117,12 +121,11 @@ describe("cloud NotebookView store bridge", () => {
       output_ids: ["out-runtime"],
     });
 
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       {
         id: "cell-1",
         cellType: "code",
         source: "print('snapshot')",
-        language: "python",
         executionId: "exec-real",
         executionCount: 3,
         outputs: [],
@@ -139,20 +142,19 @@ describe("cloud NotebookView store bridge", () => {
   });
 
   it("keeps existing output references stable when structural cloud projection adds a cell", () => {
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       codeCellWithWidgetOutput("widget-cell", "slider-model", 7),
     ]);
 
     const firstCell = getCellById("widget-cell");
     const firstOutput = getOutputById("widget-output");
 
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       codeCellWithWidgetOutput("widget-cell", "slider-model", 7),
       {
         id: "new-cell",
         cellType: "code",
         source: "",
-        language: "python",
         executionId: null,
         executionCount: null,
         outputs: [],
@@ -165,13 +167,13 @@ describe("cloud NotebookView store bridge", () => {
   });
 
   it("updates an existing output reference when the cloud output payload changes", () => {
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       codeCellWithWidgetOutput("widget-cell", "slider-model", 7),
     ]);
 
     const firstOutput = getOutputById("widget-output");
 
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       codeCellWithWidgetOutput("widget-cell", "slider-model", 8),
     ]);
 
@@ -179,7 +181,7 @@ describe("cloud NotebookView store bridge", () => {
   });
 
   it("reuses stamped output references with equal cache keys without stringifying", () => {
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       codeCellWithStampedOutput("widget-cell", "output:widget-output:1:0", "first"),
     ]);
 
@@ -187,7 +189,7 @@ describe("cloud NotebookView store bridge", () => {
     assert.ok(firstOutput);
     assert.equal("_runt_output_cache_key" in firstOutput, false);
 
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       codeCellWithStampedOutput("widget-cell", "output:widget-output:1:0", "second"),
     ]);
 
@@ -202,13 +204,13 @@ describe("cloud NotebookView store bridge", () => {
   });
 
   it("replaces stamped output references when cache keys change", () => {
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       codeCellWithStampedOutput("widget-cell", "output:widget-output:1:0", "first"),
     ]);
 
     const firstOutput = getOutputById("widget-output");
 
-    projectCloudCellsIntoNotebookViewStores([
+    projectNotebookCellsIntoViewStores([
       codeCellWithStampedOutput("widget-cell", "output:widget-output:1:1", "second"),
     ]);
 
@@ -224,35 +226,155 @@ describe("cloud NotebookView store bridge", () => {
   });
 
   it("cleans projector-owned execution and output entries for removed cells only", () => {
-    projectCloudCellsIntoNotebookViewStores([
-      codeCellWithSyntheticOutput("removed-cell", "gone"),
-      codeCellWithSyntheticOutput("kept-cell", "still here"),
+    projectNotebookCellsIntoViewStores([
+      codeCellWithIdentifiedOutput("removed-cell", "out-removed", "gone"),
+      codeCellWithIdentifiedOutput("kept-cell", "out-kept", "still here"),
     ]);
 
     assert.equal(getCellExecutionId("removed-cell"), "notebook-view-execution:removed-cell");
     assert.ok(getExecutionById("notebook-view-execution:removed-cell"));
-    assert.ok(getOutputById("notebook-view-output:removed-cell:0"));
+    assert.ok(getOutputById("out-removed"));
     assert.equal(getCellExecutionId("kept-cell"), "notebook-view-execution:kept-cell");
     assert.ok(getExecutionById("notebook-view-execution:kept-cell"));
-    assert.ok(getOutputById("notebook-view-output:kept-cell:0"));
+    assert.ok(getOutputById("out-kept"));
 
-    cleanupCloudProjectionForRemovedCells(["removed-cell"]);
+    cleanupNotebookProjectionForRemovedCells(["removed-cell"]);
 
     assert.equal(getCellExecutionId("removed-cell"), null);
     assert.equal(getExecutionById("notebook-view-execution:removed-cell"), undefined);
-    assert.equal(getOutputById("notebook-view-output:removed-cell:0"), undefined);
+    assert.equal(getOutputById("out-removed"), undefined);
     assert.equal(getCellExecutionId("kept-cell"), "notebook-view-execution:kept-cell");
     assert.ok(getExecutionById("notebook-view-execution:kept-cell"));
-    assert.ok(getOutputById("notebook-view-output:kept-cell:0"));
+    assert.ok(getOutputById("out-kept"));
+  });
+
+  it("preserves cells, outputs, and execution pointers across a same-notebook re-run", () => {
+    paintNotebook();
+    assertPainted();
+
+    const preserved = resetNotebookProjectionUnlessPreserved({
+      previousIdentity: PAINTED,
+      nextIdentity: PAINTED,
+    });
+
+    assert.equal(preserved, true);
+    assertPainted();
+  });
+
+  it("hands an instant paint to the live effect and replaces it in place", () => {
+    paintNotebook();
+    const preserved = resetNotebookProjectionUnlessPreserved({
+      previousIdentity: PAINTED,
+      nextIdentity: PAINTED,
+    });
+    assert.equal(preserved, true);
+    assertPainted();
+
+    projectNotebookCellsIntoViewStores([
+      {
+        id: "cell-code",
+        cellType: "code",
+        source: "print('live')",
+        executionId: "exec-2",
+        executionCount: 4,
+        outputs: [
+          { output_type: "stream", name: "stdout", text: "live output\n", output_id: "out-2" },
+        ],
+        metadata: {},
+      },
+    ]);
+
+    assert.deepEqual(getCellIdsSnapshot(), ["cell-code"]);
+    assert.equal(getCellExecutionId("cell-code"), "exec-2");
+    assert.equal(getOutputById("out-2")?.output_type, "stream");
+    assert.equal(getOutputById("out-1"), undefined);
+  });
+
+  it("clears every projected store on a real notebook switch", () => {
+    paintNotebook();
+
+    const preserved = resetNotebookProjectionUnlessPreserved({
+      previousIdentity: PAINTED,
+      nextIdentity: "id:nb-2",
+    });
+
+    assert.equal(preserved, false);
+    assert.deepEqual(getCellIdsSnapshot(), [] as string[]);
+    assert.equal(getCellExecutionId("cell-code"), null);
+    assert.equal(getExecutionById("exec-1"), undefined);
+    assert.equal(getOutputById("out-1"), undefined);
+  });
+
+  it("fails closed when no painted identity was recorded", () => {
+    paintNotebook();
+
+    const preserved = resetNotebookProjectionUnlessPreserved({
+      previousIdentity: null,
+      nextIdentity: PAINTED,
+    });
+
+    assert.equal(preserved, false);
+    assert.deepEqual(getCellIdsSnapshot(), [] as string[]);
+    assert.equal(getOutputById("out-1"), undefined);
+  });
+
+  it("does not preserve an empty projection", () => {
+    const preserved = resetNotebookProjectionUnlessPreserved({
+      previousIdentity: PAINTED,
+      nextIdentity: PAINTED,
+    });
+
+    assert.equal(preserved, false);
+    assert.deepEqual(getCellIdsSnapshot(), [] as string[]);
   });
 });
+
+const PAINTED = "id:nb-1";
+
+function paintNotebook(): void {
+  projectNotebookCellsIntoViewStores([
+    {
+      id: "cell-code",
+      cellType: "code",
+      source: "print('painted')",
+      executionId: "exec-1",
+      executionCount: 3,
+      outputs: [
+        {
+          output_type: "stream",
+          name: "stdout",
+          text: "painted output\n",
+          output_id: "out-1",
+        },
+      ],
+      metadata: {},
+    },
+    {
+      id: "cell-md",
+      cellType: "markdown",
+      source: "# Painted",
+      executionId: null,
+      executionCount: null,
+      outputs: [],
+      metadata: {},
+    },
+  ]);
+}
+
+function assertPainted(): void {
+  assert.deepEqual(getCellIdsSnapshot(), ["cell-code", "cell-md"]);
+  assert.equal(getCellExecutionId("cell-code"), "exec-1");
+  assert.equal(getExecutionById("exec-1")?.execution_count, 3);
+  assert.deepEqual(getExecutionById("exec-1")?.output_ids, ["out-1"]);
+  const output = getOutputById("out-1");
+  assert.equal(output?.output_type, "stream");
+}
 
 function codeCellWithWidgetOutput(cellId: string, modelId: string, value: number) {
   return {
     id: cellId,
     cellType: "code" as const,
     source: "display(slider)",
-    language: "python",
     executionId: "exec-widget",
     executionCount: 1,
     outputs: [
@@ -277,7 +399,6 @@ function codeCellWithStampedOutput(cellId: string, cacheKey: string, text: strin
     id: cellId,
     cellType: "code" as const,
     source: "display(value)",
-    language: "python",
     executionId: "exec-widget",
     executionCount: 1,
     outputs: [
@@ -296,16 +417,16 @@ function codeCellWithStampedOutput(cellId: string, cacheKey: string, text: strin
   };
 }
 
-function codeCellWithSyntheticOutput(cellId: string, text: string) {
+function codeCellWithIdentifiedOutput(cellId: string, outputId: string, text: string) {
   return {
     id: cellId,
     cellType: "code" as const,
     source: "print(value)",
-    language: "python",
     executionId: null,
     executionCount: 1,
     outputs: [
       {
+        output_id: outputId,
         output_type: "stream" as const,
         name: "stdout" as const,
         text,
