+++ src/components/notebook/capabilities.ts
@@ -7,6 +7,7 @@ export {
   projectNotebookShellCapabilities,
   projectNotebookWorkstationLaunchReadiness,
   projectNotebookWorkstationSelection,
+  projectNotebookWorkstationSurface,
   readOnlyNotebookShellCapabilities,
   resolveNotebookShellRuntimeTarget,
   stabilizeNotebookShellCapabilities,
@@ -58,9 +59,13 @@ export {
   type NotebookWorkstationPanelTone,
   type NotebookWorkstationSelectionProjection,
   type NotebookWorkstationSelectionState,
+  type NotebookWorkstationSurfaceMutationProjection,
+  type NotebookWorkstationSurfaceProjection,
+  type NotebookWorkstationToolbarActionProjection,
   type ProjectNotebookCommandRuntimeActionsOptions,
   type ProjectNotebookCommandRuntimeStatusOptions,
   type ProjectNotebookLaunchEnvironmentOptions,
   type ProjectNotebookWorkstationLaunchReadinessOptions,
   type ProjectNotebookWorkstationSelectionOptions,
+  type ProjectNotebookWorkstationSurfaceOptions,
 } from "runtimed";
